+++ app/src/main/java/org/gdg/frisbee/android/arrow/ArrowActivity.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gdg.frisbee.android.activity;
+package org.gdg.frisbee.android.arrow;
 
 import android.content.Intent;
 import android.graphics.Bitmap;
@@ -51,6 +51,7 @@ import com.google.zxing.integration.android.IntentResult;
 
 import org.gdg.frisbee.android.Const;
 import org.gdg.frisbee.android.R;
+import org.gdg.frisbee.android.activity.GdgNavDrawerActivity;
 import org.gdg.frisbee.android.app.App;
 import org.gdg.frisbee.android.app.OrganizerChecker;
 import org.gdg.frisbee.android.utils.CryptoUtils;
